+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.29.0-beta.1] (2025-12-16)
+### Changed
+- Used write and read streams to save scan responses
 
 ## [0.28.1] (2025-12-10)
 ### Changed
@@ -10,6 +13,10 @@ All notable changes to this project will be documented in this file. See [standa
 ## [0.28.0] (2025-12-10)
 ### Added
 - Added `requirement` field in dependency response
+
+## [0.28.0-beta.1] (2025-11-19)
+### Changed
+- Used write and read streams to save scan responses
 
 ## [0.27.0] (2025-11-11)
 ### Added
@@ -187,5 +194,7 @@ All notable changes to this project will be documented in this file. See [standa
 ### [0.25.0](https://github.com/scanoss/scanoss.js/compare/v0.24.0...v0.25.0) (2025-10-01)
 ### [0.26.0](https://github.com/scanoss/scanoss.js/compare/v0.25.0...v0.26.0) (2025-10-30)
 ### [0.27.0](https://github.com/scanoss/scanoss.js/compare/v0.26.0...v0.26.1) (2025-11-11)
+### [0.28.0-beta.1]  https://github.com/scanoss/scanoss.js/compare/v0.27.0...v0.28.1-beta.1 (2025-11-19)
 ### [0.28.0](https://github.com/scanoss/scanoss.js/compare/v0.27.0...v0.28.1) (2025-12-10)
 ### [0.28.1](https://github.com/scanoss/scanoss.js/compare/v0.28.0...v0.28.1) (2025-12-10)
+### [0.29.0-beta.1](https://github.com/scanoss/scanoss.js/compare/v0.28.1...v0.29.0-beta.1) (2025-12-16)